requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
